requirements.txt
selenium>=4.10
requests>=2.31
PyYAML>=6.0
